Tabs and option hyperlinks
The bulk of this is adding the new tab system that allows authors to open multiple scenes at once and quickly switch between them (DTGE-2).

Also included is the easy version of option hyperlinks, which only includes opening scenes that are already loaded into another tab or creating a new scene (DTGE-19).