Implemented scene snippets
Implemented snippets, which allows concatenation of multiple sets of text which can be switched on based on game logic. To start, only simple (one text) and random are options. This also includes some backwards compatibility in the serializer to allow scenes using the old SceneText version to be updated to snippets.