Basic Entity System
Added the components for the basic entity system, including entity setters, an if and ifelse mode for snippets, and conditionals to support those modes on snippets. Also added the ability to set "preview" values for entities and to see the end values of all entities after a scene has run.