Introduction of the SceneReadOnly and SceneEditable
This change is the minimal change to introduce the two separate wrappers around the scene to separate editing logic out from the core scene data. This is focused on the interface between the non-core namespaces and the scene, and taking full advantage of this for better encapsulation will come with follow up changes.