More snippet-related work
I adjusted the refactor to use a wrapper Snippet class around the implementations, which can hold onto the id (removing the awkward replace workaround) and all of the conversion. Ideally, the new Snippet class is the only one that needs to have big switch cases based on the snippet mode.

Since subscenes are also linked closely with snippets, I ended up doing a lot of work on them as well.