Final polish on this first wave for v0.2.0
Finished up the final touches on the first wave of this refactor. The core patterns are set with solid comments. The remainder of the refactor is cleaning up a few specific cases that need to be redone in the new pattern like null subscenes and moving options/snippets within the list.