Null subscenes redone
This commit is the new implementation of null subscenes. It introduces it as a "read only" subscene, which prevents the author from deleting it or changing the name. I'm not sure if this is something we'll use again, but I prefer having this be a more generic ability than have the concept of "null subscenes" be more explicitly shown to the editor on a per-subscene basis.